+++ src/app/authorization/login/login.component.ts
@@ -1,7 +1,7 @@
 import {Component} from '@angular/core';
 import {FormBuilder, FormGroup, Validators} from '@angular/forms';
 import {Router} from '@angular/router';
-import {AuthenticationService} from '../../authentication.service';
+import {AuthenticationService} from '../../services/authentication.service';
 import {appear, fade} from '../../animations';
 
 declare var $: any;
