--- src/app/in-memory-data.service.spec.ts
@@ -1,12 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { InMemoryDataService } from './in-memory-data.service';
-
-describe('InMemoryDataService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
-
-  it('should be created', () => {
-    const service: InMemoryDataService = TestBed.get(InMemoryDataService);
-    expect(service).toBeTruthy();
-  });
-});
